% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyOriginR-package.R
\docType{package}
\name{PolyOriginR-package}
\alias{PolyOriginR}
\alias{PolyOriginR-package}
\title{PolyOriginR: R wrapper for using PolyOrigin.jl}
\description{
Haplotype reconstruction in polyploid multiparetnal populations
}
\author{
\strong{Maintainer}: Chaozhi Zheng \email{chaozhi.zheng@wur.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
