% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyOriginR.R
\name{polyOriginR}
\alias{polyOriginR}
\title{polyOriginR}
\usage{
polyOriginR(
  genofile,
  pedfile,
  julia_home = "",
  isphysmap = FALSE,
  recomrate = 1,
  workdir = getwd(),
  outstem = "outstem",
  verbose = TRUE
)
}
\arguments{
\item{genofile}{Input genotypic data for parents and offspring}

\item{pedfile}{Input breeding pedigree}

\item{julia_home}{Path to julia.exe, Default: ''}

\item{isphysmap}{TRUE, if input marker map in genofile is physical map, Default: FALSE}

\item{recomrate}{Recombination rate in cM/Mpb, Default: 1}

\item{workdir}{Work directory, Default: getwd()}

\item{outstem}{Stem of output files, Default: 'outstem'}

\item{verbose}{TURE, if print details, Default: TRUE}
}
\value{
Return 0 if sucess, and save four output files:
log file: outstem.log,
outstem_parentphased.csv
outstem_parentphased_corrected.csv (if there exist detected errors)
outstem_genoprob.csv, and
outstem_polyancestry.csv.
}
\description{
Haplotype reconstruction in polyploid multiparental populations
}
\details{
DETAILS
PolyOriginR is implemented via PolyOriginCmd that using PolyOrigin.jl by a command line.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 polyOriginR("geno.csv","ped.csv",julia_home="C:/path/tojulia",workdir="workdir")
 }
}
}
