% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyOriginR.R
\name{polyOriginR}
\alias{polyOriginR}
\title{FUNCTION_TITLE}
\usage{
polyOriginR(
  genofile,
  pedfile,
  julia_home = "",
  isphysmap = FALSE,
  recomrate = 1,
  workdir = getwd(),
  outstem = "outstem",
  verbose = TRUE
)
}
\arguments{
\item{genofile}{PARAM_DESCRIPTION}

\item{pedfile}{PARAM_DESCRIPTION}

\item{julia_home}{PARAM_DESCRIPTION, Default: ''}

\item{isphysmap}{PARAM_DESCRIPTION, Default: FALSE}

\item{recomrate}{PARAM_DESCRIPTION, Default: 1}

\item{workdir}{PARAM_DESCRIPTION, Default: getwd()}

\item{outstem}{PARAM_DESCRIPTION, Default: 'outstem'}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
